Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -5,7 +5,7 @@
 // }
 
 int main() {
-   int a = scanf("Enter a number: ");
-   printf("a = %d", a);
+   float a = scanf("Enter a number: ");
+   printf("a = %f", a);
    return 0;
 }
